+++ src/js/components/TaskViewComponent.jsx
@@ -2,6 +2,10 @@ import lazy from "lazy.js";
 import React from "react/addons";
 
 import AppsActions from "../actions/AppsActions";
+import DialogActions from "../actions/DialogActions";
+import DialogStore from "../stores/DialogStore";
+import DialogSeverity from "../constants/DialogSeverity";
+import ExternalLinks from "../constants/ExternalLinks";
 import PagedNavComponent from "../components/PagedNavComponent";
 import TasksActions from "../actions/TasksActions";
 import TaskListComponent from "../components/TaskListComponent";
@@ -45,7 +49,7 @@ var TaskViewComponent = React.createClass({
     AppsActions.requestApp(this.props.appId);
   },
 
-  handleKillSelectedTasks: function (scaleTask) {
+  handleKillSelectedTasks: function (scaleTask, wipeTasks) {
     var props = this.props;
     var selectedTaskIds = Object.keys(this.state.selectedTasks);
 
@@ -55,11 +59,42 @@ var TaskViewComponent = React.createClass({
         : null;
     }).compact().value();
 
-    if (!scaleTask) {
-      TasksActions.deleteTasks(props.appId, taskIds);
-    } else {
+    if (scaleTask) {
       TasksActions.deleteTasksAndScale(props.appId, taskIds);
+      this.setState({selectedTasks: {}});
+      return;
     }
+
+    if (wipeTasks) {
+      const dialogId = DialogActions.confirm({
+        actionButtonLabel: "Kill and Wipe",
+        message: (
+          <div>
+            <p>This will kill all selected tasks and wipe the respective
+              local volumes. <a href={ExternalLinks.LOCAL_VOLUMES}
+                target="_blank"
+                className="modal-body-link">
+                Read more about persistent local volumes.
+              </a>
+            </p>
+            <p>
+              Are you sure you want to continue?
+            </p>
+          </div>
+        ),
+        severity: DialogSeverity.WARNING,
+        title: "Kill and Wipe Local Volumes"
+      });
+
+      DialogStore.handleUserResponse(dialogId, () => {
+        TasksActions.deleteTasks(props.appId, taskIds, true);
+        this.setState({selectedTasks: {}});
+      });
+
+      return;
+    }
+
+    TasksActions.deleteTasks(props.appId, taskIds, false);
     this.setState({selectedTasks: {}});
   },
 
@@ -115,17 +150,44 @@ var TaskViewComponent = React.createClass({
         <div className="btn-group">
           <button
               className="btn btn-sm btn-info"
-              onClick={this.handleKillSelectedTasks.bind(this, false)}>
+              onClick={this.handleKillSelectedTasks.bind(this, false, false)}>
             Kill
           </button>
           <button
               className="btn btn-sm btn-info"
-              onClick={this.handleKillSelectedTasks.bind(this, true)}>
+              onClick={this.handleKillSelectedTasks.bind(this, true, false)}>
             Kill &amp; Scale
           </button>
+          {this.getKillAndWipeButton()}
         </div>
       );
     }
+  },
+
+  getKillAndWipeButton: function () {
+    var hasStatefulTasks = Object.keys(this.state.selectedTasks)
+      .reduce((memo, taskId) => {
+        var task = this.props.tasks.find(task => task.id === taskId);
+        if (task != null &&
+            task.localVolumes != null
+            && task.localVolumes.length > 0) {
+          memo = true;
+        }
+
+        return memo;
+      }, false);
+
+    if (!hasStatefulTasks) {
+      return null;
+    }
+
+    return (
+      <button
+        className="btn btn-sm btn-info"
+        onClick={this.handleKillSelectedTasks.bind(this, false, true)}>
+        Kill &amp; Wipe
+      </button>
+    );
   },
 
   getPagedNav: function () {